foram adicionados as tabs para, disponivel, morto, excluido e vendido. Foi criado também uma tela para verificar informaçoes do caprino selecionado (ainda a melhorar)
