--- ouvinos_caprinos/lib/ui/show_caprino_information.dart
@@ -1,31 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:ouvinos_caprinos/helper/animais_helper.dart';
-
-class CaprinoInformation extends StatefulWidget {
-  final Animal caprino;
-
-  CaprinoInformation({this.caprino});
-
-  @override
-  _CaprinoInformationState createState() => _CaprinoInformationState();
-}
-
-class _CaprinoInformationState extends State<CaprinoInformation> {
-  Animal _caprinoSelecionado;
-  @override
-  void initState() {
-    super.initState();
-    _caprinoSelecionado = Animal.fromMap(widget.caprino.toMap());
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.green,
-        title: Text("Informações Sobre " + _caprinoSelecionado.nome),
-        centerTitle: true,
-      ),
-    );
-  }
-}
